+++ test-third-party-deps/testcases/jdbc_example.clj
@@ -1,0 +1,9 @@
+(ns testcases.jdbc-example
+  (:require
+   [clojure.java.jdbc :refer [with-db-connection]]
+   [clojure.test :refer [testing]]))
+
+(defn foo [x y]
+  (testing "Foo"
+    (with-db-connection [x y]
+      42)))